requirements.txt
opencv-python
streamlit
numpy
pandas
Pillow
streamlit-image-coordinates
